require mutation serialization to work via ADL
