+++ fwdpp/Makefile.in
@@ -298,7 +298,6 @@ pdfdir = @pdfdir@
 prefix = @prefix@
 program_transform_name = @program_transform_name@
 psdir = @psdir@
-runstatedir = @runstatedir@
 sbindir = @sbindir@
 sharedstatedir = @sharedstatedir@
 srcdir = @srcdir@
@@ -307,7 +306,7 @@ target_alias = @target_alias@
 top_build_prefix = @top_build_prefix@
 top_builddir = @top_builddir@
 top_srcdir = @top_srcdir@
-SUBDIRS = internal tags sugar extensions
+SUBDIRS = internal tags sugar extensions io
 pkginclude_HEADERS = debug.hpp \
 	diploid.hh \
 	fitness_models.hpp \
