+++ fwdpp/io/Makefile.am
@@ -1,0 +1,4 @@
+pkgincludedir=$(prefix)/include/fwdpp/io
+
+pkginclude_HEADERS=scalar_serialization.hpp
+
